一応動くプログラム

+++ dev_processing_systemSekkei/Arduino/speed.ino
@@ -1,7 +1,7 @@
 
 //速度を計算mm/s
 void calPos(){
-  int now_time = millis();
+  unsigned long now_time = millis();
   now_Pos[0]+=speed0*cos(DEG_TO_RAD(heading_G2))*(now_time-speed_pos_prev_time)*0.001;
   now_Pos[1]+=speed0*sin(DEG_TO_RAD(heading_G2))*(now_time-speed_pos_prev_time)*0.001;
   speed_pos_prev_time = now_time;
